+++ packages/log-http/err.js
@@ -1,11 +1,17 @@
 'use strict';
 
 module.exports = (err) => {
-    const { stack } = err;
+    const { context, stack } = err;
 
-    return {
+    const data = {
         '@type':
             'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent',
         'message': stack
     };
+
+    if (context) {
+        data.context = context;
+    }
+
+    return data;
 };