+++ packages/log-http/tests/index.test.js
@@ -61,6 +61,14 @@ const setup = (middleware) =>
             next(new Error('Testing errors...'))
         );
 
+        app.get('/error-with-context', (req, res, next) => {
+            const err = new Error('Testing errors...');
+            err.context = {
+                code: 123
+            };
+            next(err);
+        });
+
         app.get('/500', (req, res) => {
             res.statusCode = 500;
             return res.end('error');
@@ -149,7 +157,7 @@ test('logs 500 status', async (done) => {
 });
 
 test('logs errors', async (done) => {
-    expect.assertions(8);
+    expect.assertions(9);
 
     const stream = sink();
     const log = logger({ stream });
@@ -166,11 +174,39 @@ test('logs errors', async (done) => {
         expect(result.err['@type']).toBe(
             'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent'
         );
+        expect(result.err).not.toHaveProperty('context');
 
         return done();
     });
 
     get(server, '/error');
+});
+
+test('logs error context', async (done) => {
+    expect.assertions(10);
+
+    const stream = sink();
+    const log = logger({ stream });
+    const server = await setup(log);
+
+    once(stream, 'data').then((result) => {
+        expect(result).toHaveProperty('res');
+        expect(result.res).toHaveProperty('statusCode');
+        expect(result.res.statusCode).toBe(500);
+        expect(result).toHaveProperty('err');
+        expect(result.err).toHaveProperty('message');
+        expect(result.err.message).toContain('Testing errors...');
+        expect(result.err).toHaveProperty('@type');
+        expect(result.err['@type']).toBe(
+            'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent'
+        );
+        expect(result.err).toHaveProperty('context');
+        expect(result.err.context).toEqual({ code: 123 });
+
+        return done();
+    });
+
+    get(server, '/error-with-context');
 });
 
 test('logs the response content-type', async (done) => {